Revamp off2glb
